backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.5",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "numpy>=2.2.2",
    "pandas>=2.2.3",
    "rdkit>=2024.9.4",
]
